two new options: placeholder and escapeHtml
escapeHtml: if true, `text` will be used instead of `html`. This allows
for stuff such as spans and glyphicons to be used in the holder.

placeholder: if set, will be used instead of emptyMessage or when the user
hasn't selected any options. Once the user makes a selection, the holder's
text will be set to the selected item.